+++ modules/reindex/build.gradle
@@ -29,7 +29,7 @@ testClusters.configureEach {
   // Modules who's integration is explicitly tested in integration tests
   module ':modules:parent-join'
   module ':modules:lang-painless'
-  module ':modules:main'
+  module ':modules:rest-root'
   // Whitelist reindexing from the local node so we can test reindex-from-remote.
   setting 'reindex.remote.whitelist', '127.0.0.1:*'
   requiresFeature 'es.index_mode_feature_flag_registered', Version.fromString("8.0.0")
@@ -40,7 +40,7 @@ dependencies {
   api project(":libs:elasticsearch-ssl-config")
   // for parent/child testing
   testImplementation project(':modules:parent-join')
-  testImplementation project(':modules:main')
+  testImplementation project(':modules:rest-root')
 }
 
 restResources {
