+++ modules/rest-root/src/test/java/org/elasticsearch/rest/root/MainResponseTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.rest.main;
+package org.elasticsearch.rest.root;
 
 import org.elasticsearch.Build;
 import org.elasticsearch.Version;
