+++ modules/rest-root/src/yamlRestTest/java/org/elasticsearch/rest/root/RestMainClientYamlTestSuiteIT.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.rest.main;
+package org.elasticsearch.rest.root;
 
 import com.carrotsearch.randomizedtesting.annotations.Name;
 import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
