+++ modules/rest-root/src/yamlRestTest/resources/rest-api-spec/test/info/10_info.yml
