+++ modules/rest-root/src/yamlRestTest/resources/rest-api-spec/test/info/20_lucene_version.yml
