+++ modules/rest-root/src/yamlRestTest/resources/rest-api-spec/test/ping/10_ping.yml
