+++ modules/transport-netty4/build.gradle
@@ -46,7 +46,7 @@ dependencies {
   api "io.netty:netty-transport:${versions.netty}"
   api "io.netty:netty-transport-native-unix-common:${versions.netty}"
 
-  testImplementation project(':modules:main')
+  testImplementation project(':modules:rest-root')
 }
 
 restResources {
@@ -93,7 +93,7 @@ tasks.named("internalClusterTest").configure {
 }
 
 testClusters.configureEach {
-  module ':modules:main'
+  module ':modules:rest-root'
 }
 
 tasks.named("check").configure {
