+++ rest-api-spec/build.gradle
@@ -35,7 +35,7 @@ artifacts {
 
 dependencies {
   clusterModules project(":modules:mapper-extras")
-  clusterModules project(":modules:main")
+  clusterModules project(":modules:rest-root")
 }
 
 tasks.named("yamlRestTestV7CompatTransform").configure { task ->
