+++ x-pack/plugin/core/build.gradle
@@ -54,7 +54,7 @@ dependencies {
   testImplementation project(path: ':modules:parent-join')
   testImplementation project(path: ':modules:lang-mustache')
   testImplementation project(path: ':modules:analysis-common')
-  testImplementation project(path: ':modules:main')
+  testImplementation project(path: ':modules:rest-root')
   testImplementation project(":client:rest-high-level")
   // Needed for Fips140ProviderVerificationTests
   testCompileOnly('org.bouncycastle:bc-fips:1.0.2')
