+++ x-pack/plugin/security/build.gradle
@@ -26,7 +26,7 @@ dependencies {
   testImplementation project(path: ':modules:analysis-common')
   testImplementation project(path: ':modules:reindex')
   testImplementation project(':modules:data-streams')
-  testImplementation project(':modules:main')
+  testImplementation project(':modules:rest-root')
   testImplementation project(":client:rest-high-level")
 
   testImplementation(testArtifact(project(xpackModule('core'))))
