+++ x-pack/plugin/security/qa/service-account/build.gradle
@@ -6,7 +6,7 @@ dependencies {
   javaRestTestImplementation project(':client:rest-high-level')
   javaRestTestImplementation project(':x-pack:plugin:security')
   clusterModules(project(":modules:analysis-common"))
-  clusterModules(project(":modules:main"))
+  clusterModules(project(":modules:rest-root"))
 }
 
 testArtifacts {
