+++ x-pack/plugin/sql/jdbc/build.gradle
@@ -14,7 +14,7 @@ dependencies {
   api project(':x-pack:plugin:sql:sql-client')
   testImplementation project(":test:framework")
   testImplementation(testArtifact(project(xpackModule('core'))))
-  testImplementation project(':modules:main')
+  testImplementation project(':modules:rest-root')
 }
 
 tasks.named("compileJava").configure {
