+++ x-pack/plugin/sql/jdbc/src/test/java/org/elasticsearch/xpack/sql/jdbc/WebServerTestCase.java
@@ -10,7 +10,7 @@ package org.elasticsearch.xpack.sql.jdbc;
 import org.elasticsearch.Build;
 import org.elasticsearch.Version;
 import org.elasticsearch.cluster.ClusterName;
-import org.elasticsearch.rest.main.MainResponse;
+import org.elasticsearch.rest.root.MainResponse;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.test.http.MockWebServer;
 import org.junit.After;
