+++ x-pack/plugin/sql/qa/server/build.gradle
@@ -102,7 +102,7 @@ subprojects {
     javaRestTestRuntimeOnly project(path: xpackModule('spatial'))
     javaRestTestRuntimeOnly project(path: ':modules:legacy-geo')
 
-    javaRestTestRuntimeOnly project(path: ':modules:main')
+    javaRestTestRuntimeOnly project(path: ':modules:rest-root')
 
     javaRestTestRuntimeOnly "org.slf4j:slf4j-api:1.7.25"
   }
