+++ x-pack/qa/full-cluster-restart/build.gradle
@@ -11,7 +11,7 @@ dependencies {
   javaRestTestImplementation(testArtifact(project(xpackModule('core'))))
   javaRestTestImplementation(testArtifact(project(":qa:full-cluster-restart"), "javaRestTest"))
   javaRestTestImplementation project(':x-pack:qa')
-  javaRestTestImplementation project(':modules:main')
+  javaRestTestImplementation project(':modules:rest-root')
 
 }
